backend/requirements.txt
alembic==1.8.1
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.2
Flask-SSLify==0.1.5
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.43
urllib3==2.0.4
Werkzeug==2.2.2
